# bthwani-backend-nest/reports/openapi.yaml
openapi: 3.0.0
paths:
...
  /orders:
    post:
      description: إنشاء طلب جديد مع العناصر والعنوان
      operationId: OrderController_createOrder
      parameters: []
      requestBody:
        required: true
        description: بيانات الطلب
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: تم إنشاء الطلب بنجاح
        '400':
          description: بيانات غير صحيحة
        '401':
          description: غير مصرّح
      summary: إنشاء طلب جديد
      tags:
        - Order
  /orders/my-orders:
    get:
      description: جلب جميع طلبات المستخدم الحالي مع pagination
      operationId: OrderController_getMyOrders
      parameters:
        - name: cursor
          required: false
          in: query
          description: Cursor للصفحة التالية
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: 'عدد النتائج (default: 20)'
          schema:
            default: 20
            type: number
      responses:
        '200':
          description: قائمة الطلبات
        '401':
          description: غير مصرّح
      summary: جلب طلبات المستخدم
      tags:
        - Order
  /orders/{id}:
    get:
      operationId: OrderController_getOrder
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: تفاصيل الطلب
      tags:
        - Order
  /orders/{id}/assign-driver:
    post:
      description: تعيين سائق متاح للطلب (Admin/Dispatcher)
      operationId: OrderController_assignDriver
      parameters:
        - name: id
          required: true
          in: path
          description: معرّف الطلب
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                driverId:
                  type: string
                  description: معرّف السائق
              required:
                - driverId
      responses:
        '200':
          description: تم تعيين السائق بنجاح
        '401':
          description: غير مصرّح
        '404':
          description: الطلب أو السائق غير موجود
      summary: تعيين سائق للطلب
      tags:
        - Order
  /orders/{id}/notes:
    post:
      description: إضافة ملاحظة عامة أو خاصة للطلب
      operationId: OrderController_addNote
      parameters:
        - name: id
          required: true
          in: path
          description: معرّف الطلب
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  description: نص الملاحظة
              required:
                - note
      responses:
        '200':
          description: تمت إضافة الملاحظة بنجاح
        '401':
          description: غير مصرّح
        '404':
          description: الطلب غير موجود
      summary: إضافة ملاحظة للطلب
      tags:
        - Order
    get:
      operationId: OrderController_getNotes
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: جلب الملاحظات
      tags:
        - Order
  /orders/{id}/vendor-accept:
    post:
      description: قبول الطلب والبدء في التحضير
      operationId: OrderController_vendorAcceptOrder
      parameters:
        - name: id
          required: true
          in: path
          description: معرّف الطلب
          schema:
            type: string
      responses:
        '200':
          description: تم قبول الطلب بنجاح
        '400':
          description: الطلب لا يمكن قبوله
        '403':
          description: ليس لديك صلاحية
        '404':
          description: الطلب غير موجود
      summary: قبول الطلب من قبل التاجر
      tags:
        - Order
  /orders/{id}/vendor-cancel:
    post:
      description: إلغاء الطلب مع تحديد السبب
      operationId: OrderController_vendorCancelOrder
      parameters:
        - name: id
          required: true
          in: path
          description: معرّف الطلب
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: سبب الإلغاء
              required:
                - reason
      responses:
        '200':
          description: تم إلغاء الطلب بنجاح
        '403':
          description: ليس لديك صلاحية
        '404':
          description: الطلب غير موجود
      summary: إلغاء الطلب من قبل التاجر
      tags:
        - Order
  /orders/{id}/pod:
    post:
      description: إضافة صورة وتوقيع كإثبات على التسليم
      operationId: OrderController_setProofOfDelivery
      parameters:
        - name: id
          required: true
          in: path
          description: معرّف الطلب
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                imageUrl:
                  type: string
                  description: رابط صورة الإثبات
                signature:
                  type: string
                  description: توقيع العميل (اختياري)
                notes:
                  type: string
                  description: ملاحظات إضافية (اختياري)
              required:
                - imageUrl
      responses:
        '200':
          description: تم إضافة إثبات التسليم بنجاح
        '403':
          description: ليس لديك صلاحية
        '404':
          description: الطلب غير موجود
      summary: إضافة إثبات التسليم (POD)
      tags:
        - Order
    get:
      operationId: OrderController_getProofOfDelivery
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: جلب إثبات التسليم
      tags:
        - Order
  /orders/{id}/cancel:
    post:
      description: إلغاء الطلب من قبل العميل مع ذكر السبب
      operationId: OrderController_cancelOrder
      parameters:
        - name: id
          required: true
          in: path
          description: معرّف الطلب
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: سبب الإلغاء
              required:
                - reason
      responses:
        '200':
          description: تم إلغاء الطلب بنجاح
        '400':
          description: لا يمكن إلغاء الطلب
        '404':
          description: الطلب غير موجود
      summary: إلغاء الطلب
      tags:
        - Order
  /orders/{id}/return:
    post:
      description: إرجاع المنتج بعد التسليم مع ذكر السبب
      operationId: OrderController_returnOrder
      parameters:
        - name: id
          required: true
          in: path
          description: معرّف الطلب
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: سبب الإرجاع
              required:
                - reason
      responses:
        '200':
          description: تم إنشاء طلب الإرجاع بنجاح
        '400':
          description: لا يمكن إرجاع الطلب
        '404':
          description: الطلب غير موجود
      summary: طلب إرجاع المنتج
      tags:
        - Order
  /orders/{id}/rate:
    post:
      description: تقييم الطلب والخدمة بعد التسليم
      operationId: OrderController_rateOrder
      parameters:
        - name: id
          required: true
          in: path
          description: معرّف الطلب
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: number
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
              required:
                - rating
      responses:
        '200':
          description: تم التقييم بنجاح
        '400':
          description: التقييم غير صالح
        '404':
          description: الطلب غير موجود
      summary: تقييم الطلب
      tags:
        - Order
  /orders/{id}/repeat:
    post:
      description: إعادة نفس الطلب بنفس العناصر
      operationId: OrderController_repeatOrder
      parameters:
        - name: id
          required: true
          in: path
          description: معرّف الطلب السابق
          schema:
            type: string
      responses:
        '201':
          description: تم إنشاء الطلب الجديد بنجاح
        '400':
          description: لا يمكن إعادة هذا الطلب
        '404':
          description: الطلب السابق غير موجود
      summary: إعادة طلب سابق
      tags:
        - Order
  /orders/{id}/admin-status:
    patch:
      operationId: OrderController_adminChangeStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: تغيير حالة الطلب (admin)
      tags:
        - Order
  /orders/export:
    get:
      description: تصدير قائمة الطلبات بصيغة CSV أو Excel
      operationId: OrderController_exportOrders
      parameters:
        - name: startDate
          required: false
          in: query
          description: تاريخ البداية (ISO format)
          schema:
            type: string
        - name: endDate
          required: false
          in: query
          description: تاريخ النهاية (ISO format)
          schema:
            type: string
        - name: status
          required: false
          in: query
          description: حالة الطلب للتصفية
          schema:
            type: string
      responses:
        '200':
          description: ملف التصدير
        '401':
          description: غير مصرّح (admin only)
      summary: تصدير الطلبات إلى Excel/CSV
      tags:
        - Order
  /orders/{id}/tracking:
    get:
      operationId: OrderController_trackOrder
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: تتبع الطلب
      tags:
        - Order
  /orders/{id}/schedule:
    post:
      description: تحديد موعد محدد لتوصيل الطلب
      operationId: OrderController_scheduleOrder
      parameters:
        - name: id
          required: true
          in: path
          description: معرّف الطلب
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scheduledDate:
                  type: string
                  format: date-time
                  description: تاريخ ووقت التوصيل المطلوب
              required:
                - scheduledDate
      responses:
        '200':
          description: تم جدولة الطلب بنجاح
        '400':
          description: التاريخ غير صالح
        '404':
          description: الطلب غير موجود
      summary: جدولة طلب للتوصيل لاحقاً
      tags:
        - Order
  /orders/public/{id}/status:
    get:
      description: الحصول على حالة الطلب للمشاركة العامة
      operationId: OrderController_getPublicOrderStatus
      parameters:
        - name: id
          required: true
          in: path
          description: معرّف الطلب
          schema:
            type: string
      responses:
        '200':
          description: حالة الطلب
        '404':
          description: الطلب غير موجود
      summary: حالة الطلب (عام بدون مصادقة)
      tags:
        - Order
...
  /orders/{id}/driver-eta:
    get:
      operationId: OrderController_getDriverETA
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: الوقت المتوقع للوصول
      tags:
        - Order
  /orders/{id}/update-location:
    post:
      description: تحديث موقع السائق أثناء التوصيل (GPS tracking)
      operationId: OrderController_updateDriverLocation
      parameters:
        - name: id
          required: true
          in: path
          description: معرّف الطلب
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lat:
                  type: number
                  description: خط العرض
                lng:
                  type: number
                  description: خط الطول
              required:
                - lat
                - lng
      responses:
        '200':
          description: تم تحديث الموقع بنجاح
        '403':
          description: ليس لديك صلاحية (driver only)
        '404':
          description: الطلب غير موجود
      summary: تحديث موقع السائق
      tags:
        - Order
  /orders/{id}/route-history:
    get:
      operationId: OrderController_getRouteHistory
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: سجل المسار
      tags:
        - Order
...
info:
  title: Bthwani API v2
  description: NestJS API Documentation - نظام إدارة الطلبات والتجارة
  version: '2.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    OrderItemDto:
      type: object
      properties:
        productType:
          type: string
        productId:
          type: string
        name:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        store:
          type: string
        image:
          type: string
      required:
        - productType
        - productId
        - name
        - quantity
        - unitPrice
        - store
    AddressDto:
      type: object
      properties:
        label:
          type: string
        street:
          type: string
        city:
          type: string
        location:
          type: object
      required:
        - label
        - street
        - city
        - location
    CreateOrderDto:
      type: object
      properties:
        user:
          type: string
          description: معرف المستخدم
        items:
          description: عناصر الطلب
          type: array
          items:
            $ref: '#/components/schemas/OrderItemDto'
        price:
          type: number
          description: إجمالي السعر
        deliveryFee:
          type: number
          description: رسوم التوصيل
        companyShare:
          type: number
          description: حصة الشركة
        platformShare:
          type: number
          description: حصة المنصة
        address:
          description: العنوان
          allOf:
            - $ref: '#/components/schemas/AddressDto'
        paymentMethod:
          type: string
          description: طريقة الدفع
          enum:
            - cash
            - wallet
            - card
            - mixed
        orderType:
          type: string
          description: نوع الطلب
          enum:
            - marketplace
            - errand
            - utility
        walletUsed:
          type: number
          description: المبلغ المستخدم من المحفظة
        cashDue:
          type: number
          description: المبلغ المتبقي كاش
      required:
        - user
        - items
        - price
        - deliveryFee
        - companyShare
        - platformShare
        - address
        - paymentMethod
        - orderType
...